# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kampus Ariza ve Talep Sistemi API",
        "description": "Universite kampusu icin destek talep yonetim API dokumantasyonu",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tickets/{ticket}/assign": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Talebi bir personele ata",
                "operationId": "a76b0ac1e14d7e2a278b743db63469be",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atandı"
                    }
                }
            }
        },
        "/api/tickets/{ticket}/resolve": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Talebi çözüldü olarak işaretle",
                "operationId": "57ab0a53e4e51d14b59aacfbad1b08ab",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Çözüldü"
                    }
                }
            }
        },
        "api/tickets/{ticket}/close": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Talebi kapat",
                "operationId": "9f87ecf1925a8c960d4a7b28c84d3ed3",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kapatıldı"
                    }
                }
            }
        },
        "/api/tickets/{ticket}/messages": {
            "post": {
                "summary": "Talebe mesaj ekle",
                "operationId": "b9eefbbd32475bdb4f09d94da864a2f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mesaj eklendi"
                    }
                }
            }
        },
        "/api/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Talepleri listele(filtreli,sayfalı)",
                "operationId": "ddc3a5e66a83482e874c84e71ebc4505",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "priority_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "assigned_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste döndü"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Yeni talep oluştur",
                "operationId": "dfc01552ac66b35e6d09fa0f0d6dc00f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "category_id",
                                    "priority_id",
                                    "customer_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "priority_id": {
                                        "type": "integer"
                                    },
                                    "customer_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Oluşturuldu"
                    }
                }
            }
        },
        "/api/tickets/{ticket}": {
            "get": {
                "summary": "Talep detayını göster",
                "operationId": "6d7ec6f4fe57a18b4f2f799a9d74e3fb",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detay döndü"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Talebi güncelle",
                "operationId": "44d776495ca406a826940ab75987800c",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "priority_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Güncellendi"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Talebi sil",
                "operationId": "53c9d68289dd6d81f2d5fee88a363ff6",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Silindi"
                    }
                }
            }
        },
...
    },
...
}